refactor:Re-plan the payment module according to controller-service-impl standards
