+++ src/main/java/com/fitlife/payment/dto/PaymentRequest.java
@@ -1,0 +1,4 @@
+package com.fitlife.payment.dto;
+
+public class PaymentRequest {
+}
