+++ src/main/java/com/fitlife/payment/repository/PaymentRepository.java
@@ -1,5 +1,6 @@
-package com.fitlife.payment;
+package com.fitlife.payment.repository;
 
+import com.fitlife.payment.entity.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
