+++ src/main/java/com/fitlife/payment/service/impl/VnPayServiceImpl.java
@@ -1,6 +1,9 @@
-package com.fitlife.payment;
+package com.fitlife.payment.service.impl;
 
 import com.fitlife.core.config.VNPayConfig;
+import com.fitlife.payment.entity.Payment;
+import com.fitlife.payment.repository.PaymentRepository;
+import com.fitlife.payment.service.PaymentService;
 import com.fitlife.subscription.Subscription;
 import com.fitlife.subscription.SubscriptionRepository;
 import jakarta.servlet.http.HttpServletRequest;
